refactor(hir): extract shared crypto.<method> lowering helper (#1434)
The named-import path (`import { randomBytes } from "node:crypto";
randomBytes(16)`, lowered in `globals.rs`) and the dotted path
(`crypto.randomBytes(16)`, lowered in `module_static.rs`) used to
inline the same `Expr::CryptoRandom*` constructions side-by-side.
Each new crypto method added required parallel edits at both sites,
and #1419 had to keep the two lists manually in sync.

New `expr_call/crypto.rs` submodule with
`lower_crypto_passthrough(method, args)` (and a cheap
`is_passthrough_method(method)` pre-check so the caller can avoid
moving `args` into the helper when the method isn't ours). Both call
sites delegate to it.

Today's covered set: `randomBytes`, `randomUUID`, `randomFillSync`.
Methods whose lowering shape diverges between sites — dotted-only
`sha256` / `md5` (dedicated shortcuts), dotted-only `getRandomValues`
(rewrites into an instance method on the buffer), named-import-only
`createSecretKey` (synthetic PropertyGet through NativeModuleRef) —
stay at their call site. The shared helper grows naturally when a
method's shape lines up across both sites.

Pure refactor: behavior identical, verified by exercising
`randomBytes(n)` / `randomUUID()` / `randomFillSync(buf)` through
both shapes and confirming identical output.
