+++ crates/perry-hir/src/lower/expr_call/globals.rs
@@ -668,30 +668,15 @@ pub(super) fn try_global_builtins(
             // arm above (line ~3060) uses, so both call shapes
             // share one codegen path.
             if module_name == "crypto" {
+                if super::crypto::is_passthrough_method(func_name) {
+                    if let Some(expr) = super::crypto::lower_crypto_passthrough(
+                        func_name,
+                        std::mem::take(&mut args),
+                    ) {
+                        return Ok(Ok(expr));
+                    }
+                }
                 match func_name {
-                    "randomFillSync" => {
-                        if !args.is_empty() {
-                            let mut iter = args.into_iter();
-                            let buffer = iter.next().unwrap();
-                            let offset = iter.next().unwrap_or(Expr::Undefined);
-                            let size = iter.next().unwrap_or(Expr::Undefined);
-                            return Ok(Ok(Expr::CryptoRandomFillSync {
-                                buffer: Box::new(buffer),
-                                offset: Box::new(offset),
-                                size: Box::new(size),
-                            }));
-                        }
-                    }
-                    "randomUUID" => {
-                        return Ok(Ok(Expr::CryptoRandomUUID));
-                    }
-                    "randomBytes" => {
-                        if !args.is_empty() {
-                            return Ok(Ok(Expr::CryptoRandomBytes(Box::new(
-                                args.into_iter().next().unwrap(),
-                            ))));
-                        }
-                    }
                     // `createSecretKey(key, encoding?)` from a named
                     // import. Without this arm the call lowered to a
                     // generic NativeMethodCall with no dispatcher for
